% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevplus.R
\name{plot.gevplus}
\alias{plot.gevplus}
\title{Plot graphs of GEV+ fit}
\usage{
\method{plot}{gevplus}(
  x,
  q = c(2, 10, 20, 30, 50, 75, 100, 150, 200),
  ci = FALSE,
  type = c("all", "rl", "qq", "pp", "hist"),
  ...
)
}
\arguments{
\item{x}{An object of class\code{gevplus}}

\item{q}{vector of return periods, \eqn{q > 1}.}

\item{ci}{if \code{ci=TRUE}, confidence intervals will be computed.}

\item{type}{if omitted a panel with a return level plot (\code{type='rl'},
a density plot (\code{type='hist'}), a qq-plot (\code{type='qq'}) and a
probability plot (\code{tpe='pp'}) are shown.}

\item{...}{Further parameters for plotting or confidence interval calculation (alpha, R)
may also be supplied as arguments.
See e.g. \link[base]{plot}.}
}
\value{
No return value, only a plot is produced.
}
\description{
A return level plot, qq-plot, pp-plot and a histogram
with the fitted density is produced
}
\examples{
z <- extRemes::revd(100, loc = 20, scale = 10, shape = -0.2)
fit <- fgevplus(z)
plot(fit, ci = TRUE)

# Compare standard GEV with GEV+
z <- rnorm(100)
rp <- c(2, 10, 20, 30, 50, 75, 100, 150, 200)
fit_gev <- extRemes::fevd(z)
fit_gevplus <- fgevplus(z)
plot(fit_gev, rperiods = rp, type = "rl", ylim = c(0, max(z) * 1.2),
main = paste0(fit_gev$type, " / ", fit_gev$method))
plot(fit_gevplus, q = rp, ci = TRUE, type = "rl")

}
