% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevplus.R
\name{print.gevplus}
\alias{print.gevplus}
\title{Print method for object of class gevplus}
\usage{
\method{print}{gevplus}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{Object of class \code{gevplus}, fitted with GEV+.}

\item{digits}{Number of digits.}

\item{...}{Additional parameters.}
}
\value{
A nicely formatted output of the fitting results.
}
\description{
Print nicely formatted output of the fit to the GEV+
}
\examples{
z <- extRemes::revd(100, loc = 20, scale = 0.5, shape = -0.2)
fit <- fgevplus(z)
print(fit)
}
