% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevplus.R
\name{return.level.gevplus}
\alias{return.level.gevplus}
\title{Return levels and confidence intervals for the GEV+}
\usage{
\method{return.level}{gevplus}(x, return.period = c(2, 20, 100), ..., do.ci = FALSE)
}
\arguments{
\item{x}{A list as returned from function \code{\link{fgevplus}},
i.e. an object of class \code{gevplus}.}

\item{return.period}{A numeric vector of return periods.}

\item{...}{Additional parameters alpha and R for the bootstrap procedure.}

\item{do.ci}{When \code{do.ci=TRUE}, either a parametric bootstrap method is used
to calculate confidence intervals of level alpha, or the intrinsic ci functions of
package extRemes are used (see also \code{\link[extRemes]{fevd}})}
}
\value{
A tibble containing either only the estimates or additionally the
the lower and upper bounds of the confidence interval.
}
\description{
This function computes return levels and confidence intervals
for the GEV+ distribution. It handles both cases when the input fit object
is of class 'fevd' and when it's not. Confidence intervals can be constructed
with a parametric bootstrap approach.
}
\examples{
z <- extRemes::revd(100, loc = 20, scale = 0.5, shape = -0.2)
fit <- fgevplus(z)
return.level(fit)
return.level(fit, do.ci = TRUE)
return.level(fit, return.period = c(2, 10, 20, 50, 100), do.ci = TRUE)


}
