% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevplus.R
\name{return.level}
\alias{return.level}
\title{Return levels and confidence intervals for the GEV+}
\usage{
return.level(x, return.period = c(2, 20, 100), ...)
}
\arguments{
\item{x}{A list as returned from function \code{\link{fgevplus}},
i.e. an object of class \code{gevplus}.}

\item{return.period}{A numeric vector of return periods.}

\item{...}{Additional parameters alpha and R for the bootstrap procedure.}
}
\description{
This function computes return levels and confidence intervals
for the GEV+ distribution. It handles both cases when the input fit object
is of class 'fevd' and when it's not. Confidence intervals can be constructed
with a parametric bootstrap approach.
}
