% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevplus.R
\name{return.levels.gevplus}
\alias{return.levels.gevplus}
\title{Compute return levels and confidence intervals for
Generalized Extreme Value distribution with restriction to positive
shape parameters (GEV+) aka the gevplus distribution}
\usage{
return.levels.gevplus(
  x,
  return.period = c(2, 20, 100),
  do.ci = FALSE,
  alpha = 0.05,
  R = 502
)
}
\arguments{
\item{x}{A list as returned from function \code{\link{fgevplus}}.}

\item{return.period}{A numeric vector of return periods.}

\item{do.ci}{When \code{do.ci=TRUE}, either a parametric bootstrap method is used
to calculate confidence intervals of level alpha, or the intrinsic ci functions of
package extRemes are used (see also \code{\link[extRemes]{fevd}})}

\item{alpha}{A numeric value representing the significance level for
the confidence interval. Default is 0.05.}

\item{R}{An integer representing the number of bootstrap samples.
Default is 502.}
}
\value{
A tibble containing either only the estimates or additionally the
the lower and upper bounds of the confidence interval.
}
\description{
This function computes return levels and confidence intervals
for the GEV+ distribution. It handles both cases when the input fit object
is of class 'fevd' and when it's not. Confidence intervals can be constructed
with a parametric bootstrap approach.
}
\examples{
z <- extRemes::revd(100, loc = 20, scale = 0.5, shape = -0.2)
fit <- fgevplus(z)
return.levels.gevplus(fit)
return.levels.gevplus(fit, do.ci = TRUE)
return.levels.gevplus(fit, return.period = c(2, 10, 20, 50, 100), do.ci = TRUE)


}
